requirements.txt
pylast>=5.1.0
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
schedule>=1.2.0
beautifulsoup4>=4.12.0
